custom serializer for BufferedIndexDataInfo
